separate build for cli/gui
