+++ public/assets/javascript/top.js
@@ -1,11 +1,12 @@
+/* global $ */
+
 $(() => {
-  $("#login").on("click", () => {
+  $('#login').on('click', () => {
     window.electron.login();
   });
 
-
-  $("#api_login").on("click", () => {
-    const api_key = $("#api_key").val();
+  $('#api_login').on('click', () => {
+    const api_key = $('#api_key').val();
     window.electron.api_login(api_key);
   });
